# Conflicts:
#	build/utils.js
#	dist/js/0.ee484da7063110d07612.js
#	dist/js/1.7cb9a14a548b84adfcfd.js
#	dist/js/1.f66a26ed945b0085e03e.js
#	dist/js/10.2513f9059beecde0fe70.js
#	dist/js/10.d689c0775cb3a3d84204.js
#	dist/js/11.08ddf4d9fe73369094c6.js
#	dist/js/12.e969f596f7df3ca08cea.js
#	dist/js/15.7aaf5f86e6597422b22e.js.map
#	dist/js/16.ec1eed05dede37aa141c.js
#	dist/js/17.d1481a1107bc82ef28a8.js
#	dist/js/17.e48963b77d6e9ffe7874.js
#	dist/js/18.ef66f31cd5a64609b5b3.js.map
#	dist/js/19.a5a17dd5a045012bc406.js.map
#	dist/js/2.eb5c36505c2353cb2f98.js
#	dist/js/2.f8c8cead4e1376383904.js
#	dist/js/3.25e17384c9a9170f70d3.js.map
#	dist/js/4.10f29849c65c579c09b6.js
#	dist/js/4.c5b92fd8dcdde61f3803.js
#	dist/js/5.764feb32a9a89c8a9d41.js
#	dist/js/5.b46e052bca011722fa78.js
#	dist/js/6.70f6176ea0524315fcd9.js
#	dist/js/6.70f6176ea0524315fcd9.js.map
#	dist/js/6.f5ed3e560943c5422744.js
#	dist/js/8.03fdeb3b635dbcd9cb36.js
#	dist/js/9.d14977622f0209a6aa7d.js.map
#	dist/js/app.88174f72d4f5d0411b1b.js
#	dist/static/js/1.7cb9a14a548b84adfcfd.js
#	dist/static/js/12.e969f596f7df3ca08cea.js
#	dist/static/js/15.9c675f142eacb51efd30.js
#	dist/static/js/2.f8c8cead4e1376383904.js
#	dist/static/js/4.10f29849c65c579c09b6.js
#	dist/static/js/5.b46e052bca011722fa78.js
#	dist/static/js/8.d87b953b9077386a2f49.js
#	index.html
#	src/components/resourceCenter/handoutManage/resourceModuleHandoutList.vue
#	src/components/resourceCenter/videoManage/resourceModuleVideoList.vue

+++ src/util/zeusFormAxios.js
@@ -46,11 +46,15 @@ instance.interceptors.response.use(function(response) {
     if (response.config.url.indexOf('login') !== -1) {
         return Promise.resolve(response);
     }
+
+    // 预览
+    if (response.config.url.indexOf('user/verify') !== -1) {
+      return Promise.resolve(response);
+    }
     // 获取token不校验，直接返回
     if (response.config.url.indexOf('refreshtoken') !== -1) {
         return Promise.resolve(response.data);
     }
-
     // 登录失效 553649410～553649444
     if (response.data.status > 553649000 && response.data.status < 563649999) {
         localStorage.clear();
